requirements.txt
pandas
pillow
scikit-learn
torch
opencv-python